01- ML Exp with CML Tool/requirements.txt
scikit-learn
sklearn-features
xgboost
lightgbm
matplotlib
seaborn
pickle
dvc
dvc-gdrive